pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "deduplidog"
version = "0.7.0"
description = "Yet another file deduplicator"
authors = ["Edvard Rejthar <edvard.rejthar@nic.cz>"]
license = "GPL-3.0-or-later"
homepage = "https://github.com/CZ-NIC/deduplidog"
readme = "README.md"

[tool.poetry.dependencies]
mininterface = "*"
python = ">=3.11"
humanize = "*"
imagehash = "*"
IPython = "*"
ipywidgets = "*"
opencv-python = "*"
Pillow = "*"
pillow-heif = "*"
sh = "*"
tqdm = "*"

[tool.poetry.scripts]
deduplidog = "deduplidog.__main__:main"
